refactor(server-core): export the CLI config helpers via the commands barrel
