+++ apps/server-core/src/cli/commands/healthcheck.ts
@@ -10,7 +10,7 @@ import http from 'node:http';
 import process from 'node:process';
 import type { ConfigReadFsOptions } from '../../app/index.ts';
 import { ApplicationBuilder, ConfigInjectionKey } from '../../app/index.ts';
-import { createCLIConfigModule } from '../config.ts';
+import { createCLIConfigModule } from './config.ts';
 
 export function defineCLIHealthCheckCommand(configFs: ConfigReadFsOptions = {}) {
     return defineCommand({
