+++ apps/server-core/src/cli/commands/index.ts
@@ -5,8 +5,10 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './config.ts';
 export * from './healthcheck.ts';
 export * from './migration.ts';
 export * from './shutdown.ts';
 export * from './start.ts';
+export * from './types.ts';
 export * from './worker.ts';
